# openapi.yaml
openapi: 3.0.3
info:
  title: Meta Manager ALC — API esterna
  description: >
    API per integrazioni esterne di Meta Manager ALC. Espone solo l'endpoint
    di sincronizzazione manuale dei dati Meta Ads per un singolo cliente
    (stessa azione del pulsante "Aggiorna KPI" nella dashboard team).
  version: "1.0.0"
servers:
  - url: https://meta-manager-alc.vercel.app
security:
  - ApiKeyAuth: []
paths:
  /api/sync-meta:
    post:
      summary: Forza il resync di spesa/lead Meta Ads per un cliente
      description: >
        Rilegge da Meta Ads la finestra rolling degli ultimi 3 giorni per il
        cliente indicato e aggiorna (upsert) la tab MetaDaily del Google
        Sheet. Nuove campagne mai viste vengono aggiunte alla tab Campagne,
        con tipo_campagna dedotto dal prefisso tra parentesi quadre nel nome
        se riconoscibile, altrimenti vuoto.
      operationId: syncMeta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clienteId]
              properties:
                clienteId:
                  type: string
                  description: Identificativo del cliente (colonna cliente_id nella tab Clienti)
                  example: mobilieri
      responses:
        "200":
          description: Sync completato
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  righe:
                    type: integer
                    description: Numero di righe scritte/aggiornate in MetaDaily
                    example: 11
        "400":
          description: Body non valido (clienteId mancante)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errore"
        "401":
          description: API key mancante o non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errore"
        "404":
          description: Cliente non trovato o non attivo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errore"
        "502":
          description: Errore nella chiamata a Meta Ads o a Google Sheets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errore"
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: 'Header: Authorization: Bearer <API_KEY>'
  schemas:
    Errore:
      type: object
      properties:
        error:
          type: string